add doca_debug_dpdk.h to be included to another files
